+++ src/lua/shared/LogExtenderUtils.lua
@@ -26,23 +26,175 @@ LogExtenderUtils = {
         safehouse = "safehouse",
         craft = "craft",
         map_alternative = "map_alternative",
-        brush_tool = "brushtool",
+        brushtool = "brushtool",
     },
 }
 
 -- writeLog sends command to server for writting log line to file.
-LogExtenderUtils.writeLog = function(filemask, message)
+function LogExtenderUtils.writeLog(filemask, message)
     sendClientCommand("LogExtender", "write", { mask = filemask, message = message });
 end
 
 -- getLogLinePrefix generates prefix for each log lines.
 -- for ease of use, we assume that the player’s existence has been verified previously.
-LogExtenderUtils.getLogLinePrefix = function(player, action)
+function LogExtenderUtils.getLogLinePrefix(player, action)
     -- TODO: Add ownerID.
     return getCurrentUserSteamID() .. " \"" .. player:getUsername() .. "\" " .. action
 end
 
 -- getLocation returns players or vehicle location in "x,x,z" format.
-LogExtenderUtils.getLocation = function(obj)
+function LogExtenderUtils.getLocation(obj)
     return math.floor(obj:getX()) .. "," .. math.floor(obj:getY()) .. "," .. math.floor(obj:getZ());
+end
+
+-- getPlayerSafehouse iterates in server safehouse list and returns
+-- area coordinates of player's houses.
+function LogExtenderUtils.getPlayerSafehouses(player)
+    if player == nil then
+        return nil;
+    end
+
+    local safehouses = {
+        Owner = {},
+        Member = {}
+    };
+
+    local safehouseList = SafeHouse.getSafehouseList();
+    for i = 0, safehouseList:size() - 1 do
+        local safehouse = safehouseList:get(i);
+        local owner = safehouse:getOwner();
+        local members = safehouse:getPlayers();
+        local area = {
+            Top = safehouse:getX() .. "x" .. safehouse:getY(),
+            Bottom = safehouse:getX2() .. "x" .. safehouse:getY2()
+        };
+
+        if player:getUsername() == owner then
+            safehouses.Owner[#safehouses.Owner + 1] = area;
+        elseif members:size() > 0 then
+            for j = 0, members:size() - 1 do
+                if members:get(j) == player:getUsername() then
+                    safehouses.Member[#safehouses.Member + 1] = area;
+                    break;
+                end
+            end
+        end
+    end
+
+    return safehouses;
+end
+
+-- getPlayerPerks returns player perks table.
+function LogExtenderUtils.getPlayerPerks(player)
+    if player == nil then
+        return nil;
+    end
+
+    local perks = {}
+
+    for i = 0, Perks.getMaxIndex() - 1 do
+        local perk = PerkFactory.getPerk(Perks.fromIndex(i));
+
+        if perk then
+            local parent = perk:getParent();
+            if parent ~= Perks.None then
+                local perkType = tostring(perk:getType());
+                local perkLevel = player:getPerkLevel(Perks.fromIndex(i));
+                local key = "\"" .. perkType .. "\"";
+
+                table.insert(perks, key .. ":" .. perkLevel);
+            end
+        end
+    end
+
+    table.sort(perks);
+
+    return perks;
+end
+
+-- getPlayerTraits returns player traits table.
+function LogExtenderUtils.getPlayerTraits(player)
+    if player == nil then
+        return nil;
+    end
+
+    local traits = {}
+
+    for i=0, player:getTraits():size() - 1 do
+        local trait = TraitFactory.getTrait(player:getTraits():get(i));
+
+        if trait then
+            table.insert(traits, '"' .. trait:getType() .. '"');
+        end
+    end
+
+    table.sort(traits);
+
+    return traits;
+end
+
+-- getPlayerStats returns some player additional info.
+function LogExtenderUtils.getPlayerStats(player)
+    if player == nil then
+        return nil;
+    end
+
+    local stats = {}
+
+    stats.Kills = player:getZombieKills();
+    stats.Survived = math.floor(player:getHoursSurvived() * 100) / 100;
+    stats.Profession = "";
+
+    if player:getDescriptor() and player:getDescriptor():getProfession() then
+        local prof = ProfessionFactory.getProfession(player:getDescriptor():getProfession());
+        if prof then
+            stats.Profession = prof:getType();
+        end
+    end
+
+    return stats;
+end
+
+-- getPlayerHealth returns some player health information.
+function LogExtenderUtils.getPlayerHealth(player)
+    if player == nil then
+        return nil;
+    end
+
+    local bd = player:getBodyDamage()
+
+    local health = {}
+
+    health.Health = math.floor(bd:getOverallBodyHealth());
+    health.Infected = bd:IsInfected() and "true" or "false";
+
+    return health;
+end
+
+-- getVehicleInfo returns some vehicles information such as id, type and center
+-- coordinate.
+function LogExtenderUtils.getVehicleInfo(vehicle)
+    local info = {
+        ID = "0",
+        Type = "unknown",
+        Center = "10,10,0", -- Unexisting coordinate.
+    }
+
+    if vehicle == nil then
+        return info;
+    end
+
+    local id = vehicle:getID() or "0";
+    local type = "unknown";
+
+    local script = vehicle:getScript();
+    if script then
+        type = script:getName() or "unknown";
+    end;
+
+    info.ID = tostring(id);
+    info.Type = type;
+    info.Center = LogExtenderUtils.getLocation(vehicle:getCurrentSquare());
+
+    return info;
 end