+++ CHANGELOG.md
@@ -10,10 +10,12 @@ All notable changes to this project will be documented in this file.
 - Fixed error after destroy IsoRadio and IsoTelevision from Brush Tool with Sledgehammer.
 
 ### Added
-- Added brash tool logs to brashtool.txt file.
+- Added brash tool logs to `brashtool.txt` file.
 
 ### Changed
-- Marked as deprecated fuctions writeLog, getLogLinePrefix, getLocation on LogExtenderClient. This functions copied to LogExtenderUtils.
+- Marked as deprecated fuctions `writeLog`, `getLogLinePrefix`, `getLocation`, `getPlayerSafehouses`, `getPlayerPerks`, `getPlayerTraits`, `getPlayerStats`, `getPlayerHealth`, `getVehicleInfo` on LogExtenderClient. This functions copied to LogExtenderUtils.
+- Changed function `LogExtenderUtils.getPlayerSafehouses` behavior - field `Owner` now is array.
+- Changed behavior of `player.txt` log file writing - field `Owner` now is array.
 
 ## [v0.11.1] - 2022-12-22
 ### Fixed